requirements.txt
Flask==2.0.1
Flask-Login
Flask-Migrate
python-dotenv
pytest
psycopg2-binary
SQLAlchemy==1.3.8
apscheduler